[transatplan] repaired the module
The website changed a lot except for some tables.

I remade the login and added pages for functionnalities that were once on only one page.

There is a return system with a form. If you miss it, the page blocks you, even if it is a parallel one.

It seems to me that coming is missing and history for market type accounts.

The module now handles pockets.