[boursedirect] Implement transactions for "Rachat partiels"
There are at least 5 different tables with different kinds of
transactions.
Only one (versements) was scraped up to now.
Thanks to the "assert" I found several connections that crash because
they contain transactions in the other tables.
The "arbitrage" table does not contain any balance so we don't scrape
it.
This patch implements the scraping of "Rachats Partiels" transactions.