[societegenerale] Support the new virtual keyboard / new login page
The new virtual keyboard has a different look (and feel), and some very
minor adjustments are needed to handle it with current code.

This commit modifies the Captcha class to handle captchas with different
sizes, and sets the hashes for the new virtual keyboard.

This is not (yet) the use of the new VK - only prealable work so that the
new VK can be used.