[ganpatrimoine] Repair history, handle iter_coming and raise error for checking and cards history
Since I built the connector for the new website, the access to history
has changed ; now we must specify the account category otherwise we get
a 400 error.

For some abstract modules such as Groupama, the transactions for some
checking and cards are not yet available on the new website so we raise
BrowserUnavailable if the server returns a 500 error when trying to
access the account history.