[deliveroo] handle json with semicolon in it
We get an error while getting subscription of user in delivroo:

    json.decoder.JSONDecodeError: Unterminated string starting at: line 1 column 2408 (char 2407)

To get the subscription we need to get the user information (id and name)

To do this, the delivroo module, parse a json, stored in html content,
as a variable in javascript.

It parse it to get only the json using the regex :
    var ROO_PAGE_DATA = (.*?);

But it fail if there is a `;` inside the javascript.
If so, the regex will stop at the `;` and the json will be troncated

We fixed the error by enforcing the ';' at the end of the javascript
line