[ovh] Handle user-activated double factor authentication
OVH lets the user enable (or not) a 2fa, which is used during each regular login.
It's now possible to manage this 2fa with 2 new (optional) config values:
* `2fa_type` let you choose the type of 2fa (between `totp`, `sms`, `u2f`, `staticOTP`)
* `2fa_value` let you handle the value of this second factor.

You have multiple options for the value of the 2fa:
* You may let `2fa_value` absent or empty, in which case weboob should ask you
for the value to use at the time it is needed. You'll be able to use your TOTP
(mobile) application, check your received SMS, use a security code, ...
* You may also choose to use an external tool which will provide the value
(e.g. PasswordStore with OTP extension, an SMS reading application, ...)

Note: For the moment, the 'u2f' method needs specific information sent to the key
which are not available in this interface (challenge, list of registered keys).
Please do not use it at this time.

Note: OVH sometimes chokes when forms are sent too frequently. As the 2FA handling
needs to submit at least one or two forms, we slow down things a bit with
a sleep() call.