[april] New module : CapProfile + CapDocuments
This module allows you to fetch all documents related to your existing contracts.
This module is only for April Santé Prevoyance customers - there are many other
branches and contracts managed by April on other websites, but this module
only deals with https://monespace.april.fr/

Note: the current version of the website seems to list all your documents
in a single place, unrelated to your contracts. There is no way to associate
a document with one contract.
Thus only one "subscription" has been created ("documents") on which all the
documents are available.
Only when the website is changed to make the relationship between contracts
and documents visible will we be able to make the same.