[apivie] repair module and scrape new API
The module doesn't work anymore because the website changed and the code
and like super old so I rewrote most of it.
There is now an API for investments and transactions.
Also modified the browser __init__ to be able to work as a parent module
for abstract modules that have different BASEURL & APIURL.