[hsbc] Avoid duplicated deferred card transactions around debit date
To get transactions of a deferred card all history tabs links are
collected before being visited and parsed. Around debit date for the
card, the first 2 tabs might have the same transactions on them. In that
case they both bear the same html class that we can use to filter one
tab out.