[cragr] Handle redirect url in validate_new_recipient response
The redirect url was not used because it always led to the same page.
But sometimes the redirect url is the same as the current url (maybe to
display errors), so if we try to end the recipient validation, by going
to the last url, while there are errors, we get an error HTTP 404.

Closes: 64941@sibi