[cragr] Switch account_space only when needed
A number of iter_ functions use go_to_account_space() function in this
module. Sometimes using it with argument '0'  is enough and avoids
trying to switch space when not needed, especially when the next request
is retried up to 6 times upon response code 500. This commit extends that
use of the function to all accounts where no switch is needed. This
saves on requests number and avoids error 500 for users with only one
space.