[bmce] Fix FormNotFound
For bmce, child of credimutuel module on old website, we can't find a form on PorPage because it
doesn't have the 'name' argument we're looking for. It has the 'id' argument though, with the same
value 'frmMere'. PorPage on other connections of creditmutuel,
using the new website, have both id and name arguments, so this change
doesn't impact them.

Closes: 49882@sibi