[boursorama] Fix non alphanum characters virtual keyboard crash
In python3 the str.isalpha() (same for str.isalnum()) method returns True
if every symbol in the string is defined as "Letter" in the unicode
character database. It includes non "ASCII alphabetical" characters,
leading to a virtual keyboard crash.
We set a regexp in module.py to solve this issue.