[creditdunord] be compatible with children modules
creditdunord children browsers do not take the BASEURL as argument, so
creditdunord browser should do the same.

Currently there are 2 ways to use a sub-bank of creditdunord: a child module
or using the "website" choice in creditdunord. The former is favored over the
latter, but we have to stay compatible for some time. Emit a warning if the
choice option is used instead of child module.

If children modules diverge from the creditdunord module, we will have to fix
it differently than just setting the BASEURL.