weboob.browser.filters.standard: improve CleanDecimal "sign" parameter
CleanDecimal has a "sign" parameter accepting a callable, but most uses cases
are simply for forcing a sign because websites get numbers wrong. As a result,
there's frequently this contrived pasted code: "sign=lambda _: -1".
Also, this "sign" function is poorly imagined because one could pass "2" and
get a value doubled.

Allow the param to be simply an sign string (like "-"), and force the sign, to
fit the common case.