[cmso] Return empty list of recipients if no emitters matched for an account
The emitters and account do not have the same id. This force us to
create a custom id based on the account/emitter label and owner (and
co-owner if one). When there is no match between any emitter and the
account, we cannot list the recipients and make transfer from this
account.