[banquepopulaire] repair login for bpaca
"Phase" bpcesta params in authorize url is not necessary anymore for "part" bpaca user.
But for the other cases the param is still required. The phase information seems to be in
js file and the value is not hardcoded. That's why we must be vigilant if other subwebsites
migrate to this version. Moreover in bpaca the good user status is required. We can get the
information in as-ex-ano-groupe.banquepopulaire.fr API.