[caissedepargne] Manage BrowserUnavailable at login
First step for login is to submit login without password and get the
authentification method type before proceeding to it.
For some connections, response JSON returns no auth method, and an
'AUTHENTICATION_FAILED' status right at this step. While there is a
message 'Confirmez votre authentification à votre banque à distance', it
is deceitful. When attempting the connection in a navigator the same
reponse is given but a JS message 'erreur technique, Ce service est
temporairement indisponible' is displayed as well. Hence the
BrowserUnavailable error to be raised.