[creditdunord/courtois] Fix redirect blocking connections
Attempt to access to ProAccountsPage can lead instead to RedirectPage,
where we can follow a link. It leads to a code 200 page even when
bearing a 'erreur-404' in the URL.
This happens only on certain connections as a wrongly activated
security feature of the server, as discussed directly with the bank,
when there is no accounts on the aimed page.
We need to pass through not to block the connection