[sogecartenet] Improved stability of iter_history
Sometimes the search for transactions opens a popup
(even after limiting the search to one year) so now we close
it, reduce the length of the search by one month and retry.

Also force the return to the transaction search page after
selecting a card because the website sometimes does not take
us there automatically.