[hsbc] Catch BrowserUnavailable error while fetching invests
Some wealth accounts are now available only on the linebourse space. The
current way of fetching accounts results in an error message.

The linebourse space is not accessible with the current authentication
so for now we catch the error so the accounts can be fetched (although
without investments details and history).