[creditdunord/courtois] Fix faulty redirect
Attempts to access to ProAccountsPage can lead instead to RedirectPage.
It would end up to a '/sites/erreur-404' URL (but in code 200).
This happens only on certain connections, as a wrongly activated
security feature of the server, as discussed directly with the bank,
when there is no accounts on ProAccountsPage.
If redirection was followed the whole session would be broken,
without being logged out, tough. Need to log back.