[cmso] don't crash a full iter_accounts for a market_number
For some reasons, trying to go to market history during an iter_savings
can cause a ClientError. If not handled, this exception will make the
browser try to relogin and start over the iter_accounts but the session
is actualy not lost. In this case, we juste handle it when retrieving
the market order and in the worst case we won't have this information