[carrefourbanque] New implementation of CardHistoryPageJSON
The pagination of the current CardHistoryPage is not working anymore.
To get more transactions we need to use a JSON API.
The POST method need a date_recup value which is a non-determinable timestamp that we get in the html of the CardHistoryPage.
Everytime we call the JSON API, we get, the transactions of the previous calls, 40 new transactions and a new date_recup that we can use to do a new request.