[myedenred] fix 500 when fetching history
There are some accounts on which the first requests fails with a 500.
Reducing the page_size avoids these crashes.

But there are some accounts which fails on the 3rd or 4th requests,
probably because there is no more transactions to fetch.
For these ones we consider a ServerError ONLY if there was no fetched
transactions and if the error is similar.