[lcl] when some transactions details requests start failing, stop doing them
For each checking account transaction, we tried to open the transaction
dedicated page to fetch details.
Unfortunately, on some accounts, none of the transaction details page work,
we get a page telling the user is not allowed to see that info. For example,
this can happen on a professional account while the user is a private one.
They can see the list of transaction but not the full details.

We now keep the behavior but if we encountered too many failures (5) on those
details, we stop trying to go to the details.

Obnoxious validate() code is moved to a more appropriate place.