[ing] better fix for 500 errors encountered after visiting life insurances
First, don't go on a non-LoggedPage at the end of iter_accounts, it can only
lead to being delogged mindlessly.

The real problem is the site fails with a 500 when a Referer doesn not point
to /secure (what a bunch of idiots), and the life insurances URL doesn't
have /secure, so after visiting them with go(), we set a bad Referer.
Use open() instead, so the Referer is not set to a "bad" URL.