[caissedepargne] Get missing accounts from savings page
Some accounts are on the "Epargne" tab but not in the summary page where
we fetched them. We need to get the ones we missed from this page. This
also means we need to get on that tab to get history and investments.

For now we only have the issue with some life insurances so we go on the
new tab only for them, but the connector might need larger changes if we
find out that other accounts are concerned.