weboob.browser.elements: explicit params for ItemElementFromAbstractPage
When switching to regular imports, ItemElementFromAbstractPage will become
obsolete. However, they will not be removed immediately, but they will be
reimplemented to be compatible with the new system. But the reimplementation
works at import-time, not at exec-time, and it's not possible to access params
defined in outer scopes at import-time. So, we make sure the attributes are
explicitly defined, not implicitly.