[netfinca] Correct original_unitvalue & quantity for Euro funds
It seems that the display slightly changed for euro funds investments:
now they display a 'percentage' in the 'Cours' column (which usually
contains the unitvalue or the original_unitvalue) and display the
purchased amount in the 'Quantité' column.
We must ignore these values because they do not correspond to the data
usually contained in these columns.