[bred] total_offset in history to avoid infinite loop and less max requests
It seems that, in certain cases, the loop can be infinite when the
website is returning the transactions endlessly.

total_offset is here to block the while. If an account has more than
100 000 transactions, they will not have older transactions.
Because of the total_offset that previously was at 100 000 which led
at 4000 request max, I've decided to reduce the total_offset max to
50 000, halving the number of requests.