[boursorama] Fix login url
When doing a request on the login url with a `/` at the end, we receive
a 301 and are redirected to the same url without a `/` at the end. And
because of the `/` that is expected in the url, we end up not finding
the page.
Removing the `/` at the end in the regex leads to no redirection and it
matches both cases.