[boursorama] Transfer: Add support for the new pages for the new transfer wizard
Boursorama recently changed the new transfer pages on their website with
a multi steps assitant.
Add support to this new assistant to fix transfers and emitters and
recipients listing.

To be noted, the new transfer flow is now splitted into 2 different
assistants. One for immediates transfers and one for scheduled
transfers.