[cmes] New form to access investment details
We used to go to investment details using the URL attached to each
invest. All invests now have the same URL that redirects to the accounts
list. Now we must use a form to go to details.

Every invest uses the same form but with a different submit input,
so we only fetch the form once and we add the right input as we submit
it.