[creditmutuel] Fix duplicate portfolio accounts
I misunderstood the issue with duplicate IDs in the portfolio accounts
page. The issue is that liquidities accounts which were fetched on the
main accounts page are now also on the portfolio page. Because of this
change we fetched them twice.

This commit reverts the change on account IDs on the portfolio page and
ignores the liquidities accounts on this page.