[cmso] Fix find_account method for market accounts
The previous method used the account label and owner to find it on the
market accounts page. However sometimes the same owner can have two
accounts with the same label, and the matching fails. This was used to
fetch invests, history and market orders, but also to determine the
account's id.

We can get the same id from the accounts page so we do that instead and
we use it for the matching.