[creditmutuel] use new way to retrieve documents
the old page was not working but the new one is pretty close
For information:
On the website, the button makes a POST generated with javascript and a
lot of data parameters, but it is possible to use a get with a
reasonable amount of parameters.
This commit let retrieve bank statements and RIB when it is
possible to make them match with an account.
It is impossible to match for sure all accounts. I chose to retrieve
to base subscriptions on iter_account, and then get documents that can
be matched to an account. BUt some documents won't be retrieved, the
other way would have been to iter on account list given on Subscription
page and make on subscription for each, but we would lost the link
between account and subcriptions.