[lcl] Fixes recipient_iban changed during transfer processing AssertionError
Some recipients ibans cannot be found anymore on the website.
But since we kept the iban stored on our side, the 'old'
transfer.recipient_iban is not empty when making a transfer.

For example, this could be the case for internal accounts like "Livret
A".
In such a case, there encountered exception was like:

AssertionError: recipient_iban changed during transfer processing (from
"FR01234XXXXXXXXXXXXXXX42" to "NotAvailable")

Closes: 168427@sibi