[creditmutuel] init_login: ensure at least 2 redirections after the login to reach the validation.aspx page
After login, we don't redirect automatically on 302 but only redirect
once if not to "validation.aspx" or that we are in interactive mode.

The reason for that is that it is possible that a sca will be triggered
when visiting this page. And so, a sms or an app notification would be
sent unexpectedly to the user.

This issue is that a redirect to this page could also be required by the
bank around 1 month before the end of the 90 days to show a warning that
sca will have to be done again soon.
Even if it will still be possible to say "skip the sca for now".

In the current case, the issue is that we were checking for the
mobile_confirmation page only after a single redirect, but sometimes
will show up as the second redirect after authentication.
Ex.: authentification.html (302) -> pageaccueil.aspx (302) ->
validation.aspx