[boursorama] Add support for EstimateFees page in the transfer wizard
In some cases, there could be an EstimateFees page that could appear in
the transfer wizard before the transfer confirmation page.

I encountered this case with a special "CEFP" account that shows this
page for giving the estimation of the estimated annual "management" fees
for the additional amount, despite the transfer fee still being of 0
euro.