[bp] Refactor handling of loi6902 popup during a transfer execution
Loi 6902 forbid banks to allow "Livret" accounts to be used for
"paiments" / "prélévements" or any transfer to an account that not a
debit account with the same owner.

More than that, most banks consider that the recipient debit account
should be located in the same bank.

A lot of banks were not respecting this law and last year there was a
crackdown on this subject by authorities.

So, before, the banque postale was allowing a transfer to external
account providing that the user would attest that he is the owner of the
external account.

But, it looks like that there was an update of the BP website around the
24/11/2020, and that now, such transfers are simply completely blocked
without possibility to ignore the error.