[fortuneo] Fix: force trigger 2FA in init_login if it is needed.
By default, after login, we are redirected to the first accounts_page URL.
It will bear a basic list of accounts if 2FA is already done.
But if it is needed, but bank doesn't have a way to notify user,
then this page will be the same list.
twofa_page will nevertherless show up later when requesting accounts details.
So we now force go to another accounts_page URL to trigger it during loging process.