weboob.tools.backend: Fixes AbstractModule CONFIG with multiple inheritance
In the case of an AbstractModule A depending on another AbstractModule
B.
B.CONFIG could be broken if A is loaded before B, for example because an
operation with A is requested first after weboob startup.

The root cause of the issue is that, to load A, the call to its
"__new__" function will rewrite the parents of B on the process but will
not take care to setup it's CONFIG correctly.

Then, if you try to create an instance of B to use it directly, it will
not be considered as an AbstractModule and so it will not be initialized
with AbstractModule.__new__() to setup its CONFIG.

As an example, there is the case of the "caissedepargne" module that
depends on "banquepopulaire", that itself depends on "stet".

This change ensure that, if needed, any intermediary dependency is
properly loaded.