[bred] Better handling of new recipient transfer limit
With a request with a limit way too high, we can find the maximum
allowed transfer limit for new recipients. This allow us to use
this limit instead of asking the user the limit they want for
this new recipient. This transfer limit can be changed on the website by the
user.