[bp] don't raise BrowserUnavailable when login is successful
The pro website has changed, and an url '.*voscomptes/identification/identification.ea.*'
that was once handled as a temporary unvailability page is now an
expected page of the login process.

I don't know if this url has to be treated the same way or differently for
the par website, so I decided to override the parent attribute in the
subclass BProBrowser.

We can also notice the following requests don't work because the pro website
has changed (a wrong request will typically lead to a deconnection, for
ex. the next_request auth_page() fails this way), but not for all
connections (strange), meaning for some connections, the current
requests work fine. So far I have noticed the current code still works
for connections from users who have already validated the SCA (but I
might have missed something else).