[bp] rework pro website
Only for "professionnels" connections (current BProBrowser). It is the
first part of the rework, other changes will come next. What is
addressed so far:

- login without sca, accounts, history, profile

By the way, it also has the benefit of removing dead code: the variable
`accounts_and_loans_url` doesn't seem to be used

TODO:

- implement the SCA

- check if there are changes for wealth and bill with appropriate
connections, and do the changes if needed

- add new account types, etc...