[caissedepargne] Handle Multi Spaces
Here the difficulty is to know which spaces to use with just the nuser
for information. If no nuser is given we use part, else we use the
second space. AssertionError has been added in case we have more than 2
spaces and 2 spaces whitout part. I don't know if these case can exist,
but we can catch it first.