[cragr] don't crash if subscription is found multiple times
This can happen because it is displayed twice (bank statement + deffered
statements). But the attributes of the subscription were the same.
Before this changes, the duplicates were supposed to be handled by the
variable parsed_subscription_ids but its values were given only at the
first iteration, so it was just an empty list and the condition was
always True.
Since the regexp which get some information on the subscription is now
used once per subscription, there were no need to make a method of it.