weboob.browser.browser.Browser.raise_for_status : raise exceptions where they are
Previously exceptions were all raised on the same line. This breaks external monitoring tools like sentry
that aggregate exceptions by stacktrace.

Because all those exceptions ARE raised at the same line in the code, those tools aggregate them.

The solution here is to raise the exception directly when we decide what type of exception it is.

The code is also easier to read if you raise the exception directly when you find it.