browser.filters.json: remove confusing Dict class __getitem__
A metaclass was used with the Dict filter to be able to use it like that:
Dict['hello'] instead of Dict('hello')

It looks like to not be used.

Thus, the code could be simplified, and "special" undocumented usages
avoided. And so this filter will also not have a specificity compared to
other filters.