[societegenerale] Fix: handle Secure Access on Ent
Secure Access is a kind of AppVal and is the 2FA method for Ent users.
It was implemented with the common SocieteGeneraleTwoFactorBrowser.

But it a systematic 2FA and
it could not work with the backend for several reasons:
- when entering the handle_polling,
first URL to be called, when loading state, was initiating a second 2FA;
- only one user's device name was displayed in the validation prompt,
although there could be several;
- after validation, we must to end up on a LoggedPage to
avoid triggering a chained init_login();
- once in background sync, load_state was starting at last url,
altough it needs to start at login to be able to raise NeedInteractiveFor2FA;
- also, since SGProfessionalBrowser is inheriting from
SGEnterpriseBrowser, we must be revert in it some of the discussed
changes made in SGEnterpriseBrowser.