[boursorama] Correct investment labels & codes
For the labels, we are fetching an <a> tag, except that some invests
have no link so we were retrieving invests with empty labels.
We now handle <a> or <strong> cases with a Coalesce.
Codes were handled with IsinCode & code_types with IsinType.
Used Base filter instead of ugly filter()(self)[0].xpath().