[caissedepargne] Fetch market accounts and manage linebourse space
Market accounts are situated on a specific page that is identical to the
accounts page. The details of some of them can be fetched by going to
the linebourse space. Unfortunately, not all of them are present, and we
need to go a first time to see which accounts are available on
linebourse so we don't go there for nothing.