[bred] Let the user choose the order of SCA methods
The default SCA methods were : sms, then app notification, then otp.
This patch allows the user to configure (in config file), using the preferred_sca
key, the list of authentication methods he wants to allow (for example, to have
SMS first, then otp, then App notification, etc...)

As for now, the following methods are supported by the module:
* notification (mobile application notification)
* sms (sms on your trusted mobile phone)
* otp (e-Code that must be generated on the mobile application)

device name (which appears in the website) can be customized with the configuration
file.