[societegenerale] Big Pro/Ent cleanup
To better avoid problems when changing things not common
to SGEntreprise and SGProfessional,
inheritance scheme is changed.

A common SGPEBrowser still exists. It takes in
common login steps from SocieteGeneraleLogin,
and common Pro/Ent navigation.
But SGProfessionalBrowser and SGEnterpriseBrowser are now
each inheriting separately from SGPEBrowser.

It bears importance for the few differences in 2FA handling
for Ent, where it is systematic, compared to Pro.

Also, dead code is deleted (browser and pages' side).
And a few commenting is added.