[kraken] Handle some API errors
The "Rate limit exceeded" error was raised as a BrowserUnavailable.
Instead we can wait for the calls counter to reset and resume browsing.

If the API key given by the user lacks permissions for what we're doing
(like fetching transactions), we get the error "Permission denied", we
raise it as an ActionNeeded so the user can configure the key.

The error message for an incorred API key can be different.