[bnporc] Fixed new recipient to be more resilient and be able to recover from a previous attempt
When the otp step was not completed or failed, a recipient was still
created on the bnp website with the "to be activated state".

It was not possible to do more than a single otp code attempt without
going into an unexpected state.
And trying to add again the recipient would fail with the following
error:
"Bénéficiaire en doublon. Code bic / iban / compte déjà présent en
base."

This change will detect if the recipient already exists in a "to be
validated" state, and in that case will just trigger the activation otp
or decoupled instead of trying to add the recipient again.