[creditdunord] New version of the module using JSON pages
There was an issue in the iter_accounts of kolb, a child of this
connector, due to some pages now loading with javascript which broke the
navigation.

Instead of adding messy ways of working around the issue in an already
messy connector I remade it using the much cleaner new JSON pages of the
"Synthèse Multibanque" space.

However, there is no way to get the same IDs as the old version and
accounts fetched with this new version don't match the old ones.