[societegenerale] Better detect logged out state
When reloading state and requesting the last page visited,
when that page is one of the one in json_pages.py,
we are usally logged out.
This is handled in AccountJsonPage with a special Logged property.
But it needs to be as well in the other pages.
We do that by introducing a LoggedDetectionMixin,
inherited by a SGPEJsonPage, common to all subsequent pages.