[cmes] Fix "investment details" form
When we use the same form for all investments, even if we clean it each
time, we arrive on the same page (the first one we got to using the form),
so all investments and pockets had the same additional info, like quantity
for example. We don't have the issue when we re-fetch the form each
time.