[boursorama] Add new XPath to check if transaction is split
Split sub-transactions are filtered out by `condition`. However the current XPath
to get the `obj_amount`, matches multiple element (the global transaction
and the split sub-transactions) causing an AssertionError using
the `CleanDecimal.French` filter.

To solve this, we first check if the transaction is split. If it is we
only use the first matched element (which corresponds to the
transaction's total amount)