woob: introduce package, importing weboob for now
After years of controversy around the project name and content because of
poor-taste jokes, the project and its apps will be renamed and the
offensive content be removed. We are sorry for the past offensive content
and we apologize for any harm that has been caused by it.

In the next release, everything will be in the woob namespace. In the
meantime, we keep both packages, so apps can test the transition.
This is the first step of many towards full renaming of the project.