[hsbc] Handle the fingerprint form page (lifeinsurance)
... that could appear when going on some lifeinsurances accounts

When we are trying to go on some lifeinsurances account pages,
sometimes we get instead an intermediary blank page with a form.

It looks like that a javascript browser fingerprint has to be done to
fill the form and submit it.
Otherwise, when we submit without it or a wrong value, we are redirected
back to the home page.

I'm not completely sure why it is sometimes needed, and if ever we would
go to the right page if ever the fingerprint was correct.

I guess that might be the case for empty or special state life
insurances.

Anyway, for the moment, we just properly detect that we arrived on this
page and five up for this life insurance account.